Fix small mistakes

+++ src/Controller/Back/PrestationController.php
@@ -13,20 +13,20 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route('/prestation')]
-#[Security('is_granted("ROLE_ADMIN") or (is_granted("ROLE_USER"))') ]
+#[Security('is_granted("ROLE_ADMIN") or (is_granted("ROLE_USER"))')]
 class PrestationController extends AbstractController
 {
     #[Route('/', name: 'app_prestation_index', methods: ['GET'])]
-    public function index(Request $request,PrestationRepository $prestationRepository): Response
+    public function index(Request $request, PrestationRepository $prestationRepository): Response
     {
-        $prestations = $prestationRepository->findAll();
-        $prestation = new Prestation();
-        $form = $this->createForm(PrestationType::class, $prestation);
-        $form->handleRequest($request);
+        if (in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            $prestations = $prestationRepository->findAll();
+        } else {
+            $prestations = $prestationRepository->findAllPrestations($this->getUser()->getCompany()->getId());
+        }
         return $this->render('prestation/index.html.twig', [
-            'prestations' => $prestationRepository->findAll(),
-            'form' => $form,
-
+            'prestations' => $prestations,
+            'connectedUser' => $this->getUser()
         ]);
     }
 
@@ -43,6 +43,7 @@ class PrestationController extends AbstractController
                 $this->addFlash('danger', 'Cette prestation existe déjà');
                 return $this->redirectToRoute('app_prestation_index');
             }
+            $prestation->setCompany($this->getUser()->getCompany());
             $entityManager->persist($prestation);
             $entityManager->flush();
             $this->addFlash('success', 'Prestation ajoutée avec succès');
@@ -52,6 +53,7 @@ class PrestationController extends AbstractController
         return $this->render('prestation/new.html.twig', [
             'prestation' => $prestation,
             'form' => $form,
+            'connectedUser' => $this->getUser()
         ]);
     }
 
@@ -61,40 +63,42 @@ class PrestationController extends AbstractController
     {
         return $this->render('prestation/show.html.twig', [
             'prestation' => $prestation,
+            'connectedUser' => $this->getUser()
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_prestation_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Prestation $prestation, EntityManagerInterface $entityManager): Response
-    {if (count($prestation->getInvoicePrestations()) > 0) {
-        $newPrestation = clone $prestation;
-        $entityManager->detach($newPrestation);
-        $prestationToPersist = $newPrestation;
-        $prestationToPersist->setArchive(false);
-        $prestation->setArchive(true);
-        $entityManager->persist($prestation);
-        $this->addFlash('warning', 'Attention, cette prestation est déjà liée à une facture, vous ne pouvez pas la modifier');
-    } else {
-        $prestationToPersist = $prestation;
-    }
+    {
+        if (count($prestation->getInvoicePrestations()) > 0) {
+            $newPrestation = clone $prestation;
+            $entityManager->detach($newPrestation);
+            $prestationToPersist = $newPrestation;
+            $prestationToPersist->setArchive(false);
+            $prestation->setArchive(true);
+            $entityManager->persist($prestation);
+            $this->addFlash('warning', 'Attention, cette prestation est déjà liée à une facture, vous ne pouvez pas la modifier');
+        } else {
+            $prestationToPersist = $prestation;
+        }
         $form = $this->createForm(PrestationType::class, $prestationToPersist);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($prestationToPersist);
             $entityManager->flush();
             $this->addFlash('success', 'Prestation modifiée avec succès');
-
         }
         return $this->render('prestation/edit.html.twig', [
             'prestation' => $prestationToPersist,
             'form' => $form,
+            'connectedUser' => $this->getUser()
         ]);
     }
 
     #[Route('/{id}', name: 'app_prestation_delete', methods: ['POST'])]
     public function delete(Request $request, Prestation $prestation, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$prestation->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete' . $prestation->getId(), $request->request->get('_token'))) {
             $invoicePrestations = $prestation->getInvoicePrestations();
             foreach ($invoicePrestations as $invoicePrestation) {
                 $entityManager->remove($invoicePrestation);
